refactor: pagination types
