+++ projects/ion/src/lib/table/table.component.ts
@@ -1,8 +1,8 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { PageEvent } from '../pagination/pagination.component';
+import { CheckBoxStates } from '../core/types/checkbox';
+import { PageEvent } from '../core/types/pagination';
 import { SafeAny } from '../utils/safe-any';
 import { ActionTable, Column, ConfigTable, TableUtils } from './utilsTable';
-import { CheckBoxStates } from '../core/types/checkbox';
 
 interface TableEvent {
   rows_selected: SafeAny[];
